HAT/backend/requirements.txt
beautifulsoup4==4.6.3
chardet==3.0.4
click==6.7
filelock==3.0.4
Flask==0.12.2
Flask-Cors==3.0.3
google==2.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
joblib==0.13.2
lxml==4.2.1
MarkupSafe==1.0
nltk==3.3
numpy==1.15.1
pandas==0.23.4
protobuf==3.6.1
pymongo==3.6.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.18.4
scikit-learn==0.21.2
scipy==1.3.0
six==1.11.0
sklearn==0.0
urllib3==1.22
uWSGI==2.0.17
Werkzeug==0.14.1
xlrd==1.1.0
